Move the PMIxUnit test runner to the TestRun Stage
Allow specification of the unit test cmd line in a one-test-at-a-time
format and use custom code to parse the input into a list of tests to
execute.
